[BE] ADD : weather api

See merge request s10-s-project/S10P21S006!63

+++ BE/smartclothing/src/main/java/sueprtizen/smartclothing/domain/outfit/recommended/dto/WeatherDTO.java
@@ -5,7 +5,7 @@ import lombok.Builder;
 @Builder
 public record WeatherDTO(
         int icon,
-        int lowestTemperature,
-        int highestTemperature
+        double lowestTemperature,
+        double highestTemperature
 ) {
 }
